fix: preserve meaningful coverage boundaries
