+++ AGENTS.md
@@ -40,6 +40,7 @@ Build for longevity. Favor readable, skimmable, well-verified code over speed or
 - The root exposes the Effect-patched native TypeScript 7 compiler as `tsc`. The `typescript` package remains the TypeScript 6 compatibility API for Next.js, Ultracite, and language-service consumers. `packages/backend` owns its package-local native compiler because Convex resolves it directly. Verify with `pnpm exec tsc --version`; use `pnpm exec tsc6 --version` only for compatibility diagnostics.
 - Formatting is owned by Ultracite. Use spaces, double quotes, `import type`, and clear external, workspace, then app-local import groups. Run `pnpm format` instead of hand-formatting.
 - New or touched app TypeScript modules use direct `@/` imports for same-app modules, including colocated modules and tests. Across workspaces use `@repo/*`. Prefer direct owning-file imports over new barrels.
+- New hand-written filenames use one domain word plus conventional suffixes such as `.client` or `.test`. Do not introduce hyphenated compound basenames.
 - Keep Tailwind class strings inside styling utilities or component boundaries. Use `cva` or existing variant helpers for reusable or variant-driven styling.
 
 ## Effect V4 Standard
@@ -91,6 +92,8 @@ Build for longevity. Favor readable, skimmable, well-verified code over speed or
 - Do not create a test merely because a `.ts` file exists, to mirror implementation details, or to satisfy coverage. Every test must prove meaningful behavior, a regression, or a failure contract at the owning public seam. Delete tests whose only value is restating configuration or exercising trivial branches. Maintain 100% statement, branch, function, and line coverage without lowering thresholds or excluding behavior. Keep declarative files with no executable behavior outside the coverage surface instead of manufacturing a test.
 - Import test APIs from `@effect/vitest`. Use the shared configured `vi` global for mocks because Vitest hoists mock calls before re-export bindings initialize. Do not import `vi`. Keep `vitest` installed only because `@effect/vitest`, the CLI runner, coverage, and `vitest/config` require it. Raw `vitest` imports are forbidden in authored TypeScript.
 - Do not add React component tests that mock children to verify static markup. Move testable behavior into an owning `.ts` domain seam and verify rendered behavior through production-mode Browser or E2E acceptance.
+- Never create a `.ts` module or unit test only to satisfy test-name checks. Do not unit-test `.tsx` modules. A domain seam must have independent production value; otherwise verify the real rendered behavior through Browser, E2E, or an existing production-boundary integration test.
+- Never remove or relax a coverage gate to accommodate a design change. Keep meaningful `.ts` seams at 100% per-file coverage and verify `.tsx` output through Browser, E2E, or production-boundary integration tests.
 - Keep tests behavior-oriented, focused, and free of `.only` or `.skip`. Run the nearest test first, then the relevant workspace suite when risk warrants it. Preserve every workspace's configured per-file 100% coverage gate.
 - Use `pnpm run doctor --verbose --scope changed --base main --include-untracked` for changed React code and `pnpm run doctor --verbose --scope full` for a whole-codebase audit. Do not use the deprecated `--diff` alias or plain `npx react-doctor@latest`.
 - Authored content follows the audited Aksara locale equivalent without fallback. Preserve reviewed facts, pedagogy, exercises, renderer contracts, and the language being assessed.
