+++ apps/www/app/[locale]/(entry)/onboarding/page.tsx
@@ -4,7 +4,7 @@ import type { Metadata } from "next";
 import { notFound } from "next/navigation";
 import { getTranslations } from "next-intl/server";
 import { Suspense } from "react";
-import { AdmissionPending } from "@/components/auth/admission-pending";
+import { AdmissionPending } from "@/components/auth/pending";
 import { Theme } from "@/components/marketing/shared/footer-action";
 import { OnboardingQuestionnaire } from "@/components/programs/onboarding/questionnaire";
 import {
