+++ packages/backend/convex/_generated/api.d.ts
@@ -459,7 +459,7 @@ import type * as routes_agent_search from "../routes/agent/search.js";
 import type * as routes_agent_security from "../routes/agent/security.js";
 import type * as routes_constants from "../routes/constants.js";
 import type * as routes_middleware_requestId from "../routes/middleware/requestId.js";
-import type * as routes_middleware_requestLogger from "../routes/middleware/requestLogger.js";
+import type * as routes_middleware_logger from "../routes/middleware/logger.js";
 import type * as routes_polar from "../routes/polar.js";
 import type * as routes_resend from "../routes/resend.js";
 import type * as schools_mutations from "../schools/mutations.js";
@@ -1050,7 +1050,7 @@ declare const fullApi: ApiFromModules<{
   "routes/agent/security": typeof routes_agent_security;
   "routes/constants": typeof routes_constants;
   "routes/middleware/requestId": typeof routes_middleware_requestId;
-  "routes/middleware/requestLogger": typeof routes_middleware_requestLogger;
+  "routes/middleware/logger": typeof routes_middleware_logger;
   "routes/polar": typeof routes_polar;
   "routes/resend": typeof routes_resend;
   "schools/mutations": typeof schools_mutations;
